Clean up shell script
There were a couple latent problems with spaces in the inputs, but most
importantly if the go script failed, this script would not fail, and
thus that failure wouldn't be passed back through to Actions.

We should probably remove the shell script entirely and just have the go
script output things in the right format, but this is the first easy
change.